+++ packages/gatsby/src/bootstrap/index.js
@@ -12,6 +12,8 @@ const telemetry = require(`gatsby-telemetry`)
 
 const apiRunnerNode = require(`../utils/api-runner-node`)
 import { getBrowsersList } from "../utils/browserslist"
+import { createSchemaCustomization } from "../utils/create-schema-customization"
+import { startPluginRunner } from "../redux/plugin-runner"
 const { store, emitter } = require(`../redux`)
 const loadPlugins = require(`./load-plugins`)
 const loadThemes = require(`./load-themes`)
@@ -20,8 +22,6 @@ const getConfigFile = require(`./get-config-file`)
 const tracer = require(`opentracing`).globalTracer()
 const preferDefault = require(`./prefer-default`)
 const removeStaleJobs = require(`./remove-stale-jobs`)
-// Add `util.promisify` polyfill for old node versions
-require(`util.promisify/shim`)()
 
 // Show stack trace on unhandled promises.
 process.on(`unhandledRejection`, (reason, p) => {
@@ -67,7 +67,7 @@ module.exports = async (args: BootstrapArgs) => {
 
   // Start plugin runner which listens to the store
   // and invokes Gatsby API based on actions.
-  require(`../redux/plugin-runner`)
+  startPluginRunner()
 
   const directory = slash(args.directory)
 
@@ -434,7 +434,7 @@ module.exports = async (args: BootstrapArgs) => {
     parentSpan: bootstrapSpan,
   })
   activity.start()
-  await require(`../utils/create-schema-customization`)({
+  await createSchemaCustomization({
     parentSpan: bootstrapSpan,
   })
   activity.end()