+++ packages/gatsby/src/query/__tests__/data-tracking.js
@@ -430,7 +430,7 @@ describe(`query caching between builds`, () => {
             siteMetadata: {
               title: `My Site`,
               description: `Description of site
-            
+
               --edited
               edited content #${nodeChangeCounter++}
               `,
@@ -509,7 +509,7 @@ describe(`query caching between builds`, () => {
             id: `test-2`,
             slug: `bar`,
             content: `Dolor sit amet.
-            
+
             --edited
             edited content #${nodeChangeCounter++}
             `,
@@ -607,7 +607,7 @@ describe(`query caching between builds`, () => {
           createNotUsedNode({
             id: `not-used`,
             content: `Content
-            
+
             --edited
             edited content #${nodeChangeCounter++}`,
           })
@@ -679,7 +679,7 @@ describe(`query caching between builds`, () => {
             id: `test-1`,
             slug: `foo`,
             content: `Lorem ipsum.
-            
+
             --edited
             edited content #${nodeChangeCounter++}`,
           })
@@ -838,7 +838,7 @@ describe(`query caching between builds`, () => {
             siteMetadata: {
               title: `My Site`,
               description: `Description of site
-            
+
               --edited
               edited content #${nodeChangeCounter++}
               `,