+++ packages/gatsby/src/redux/actions/public.ts
@@ -26,8 +26,7 @@ import { getNonGatsbyCodeFrame } from "../../utils/stack-trace-utils"
 import {
   IGatsbyPlugin,
   IPageInput,
-  IReduxNode,
-  IPage,
+  IGatsbyPage,
   IJob,
   IJobV2,
   IPageDataRemove,
@@ -57,6 +56,7 @@ import {
   ICreatePageDependencyAction,
   ISetPageDataAction,
   IRemovePageDataAction,
+  IGatsbyNode,
 } from "../types"
 import webpack from "webpack"
 
@@ -86,8 +86,8 @@ const ensureWindowsDriveIsUppercase = (filePath: string): string => {
 }
 
 const findChildren = (
-  initialChildren: IReduxNode["id"][]
-): IReduxNode["id"][] => {
+  initialChildren: IGatsbyNode["id"][]
+): IGatsbyNode["id"][] => {
   const children = [...initialChildren]
   const queue = [...initialChildren]
   const traversedNodes = new Set()
@@ -365,7 +365,7 @@ ${reservedFields.map(f => `  * "${f}"`).join(`\n`)}
     page.path = truncatedPath
   }
 
-  const internalPage: IPage = {
+  const internalPage: IGatsbyPage = {
     internalComponentName,
     path: page.path,
     matchPath: page.matchPath,
@@ -471,7 +471,7 @@ ${reservedFields.map(f => `  * "${f}"`).join(`\n`)}
  * Delete a node
  */
 export const deleteNode = (
-  options: IReduxNode["id"] | { node: IReduxNode },
+  options: IGatsbyNode["id"] | { node: IGatsbyNode },
   plugin: IGatsbyPlugin,
   args: IGatsbyPlugin
 ): IDeleteNodeAction | IDeleteNodeAction[] => {
@@ -497,7 +497,7 @@ export const deleteNode = (
 
   // Always get node from the store, as the node we get as an arg
   // might already have been deleted.
-  const node = getNode(id) as IReduxNode
+  const node = getNode(id) as IGatsbyNode
   if (plugin) {
     const pluginName = plugin.name
 
@@ -521,7 +521,7 @@ export const deleteNode = (
     /* eslint-enable @typescript-eslint/no-use-before-define */
   }
 
-  const createDeleteAction = (node: IReduxNode): IDeleteNodeAction => {
+  const createDeleteAction = (node: IGatsbyNode): IDeleteNodeAction => {
     return {
       type: `DELETE_NODE`,
       plugin,
@@ -536,7 +536,7 @@ export const deleteNode = (
   const deleteDescendantsActions =
     node &&
     findChildren(node.children)
-      .map<IReduxNode>(getNode)
+      .map<IGatsbyNode>(getNode)
       .map(createDeleteAction)
 
   if (deleteDescendantsActions && deleteDescendantsActions.length) {
@@ -574,7 +574,7 @@ export const deleteNodes = (
     // Payload contains node IDs but inference-metadata and loki reducers require
     // full node instances
     payload: nodeIds,
-    fullNodes: nodeIds.map(getNode) as IReduxNode[],
+    fullNodes: nodeIds.map(getNode) as IGatsbyNode[],
   }
 }
 
@@ -588,7 +588,7 @@ const typeOwners = {}
  * Create a new node.
  */
 type CreateNode = (
-  node: IReduxNode,
+  node: IGatsbyNode,
   plugin?: IGatsbyPlugin,
   actionOptions?: IActionOptions
 ) =>
@@ -609,7 +609,7 @@ const _createNode: CreateNode = (node, plugin, actionOptions) => {
 
   // Ensure the new node has an internals object.
   if (!node.internal) {
-    node.internal = {} as IReduxNode["internal"]
+    node.internal = {} as IGatsbyNode["internal"]
   }
 
   NODE_COUNTER++
@@ -697,7 +697,7 @@ const _createNode: CreateNode = (node, plugin, actionOptions) => {
 
   node = sanitizeNode(node)
 
-  const oldNode = getNode(node.id) as IReduxNode
+  const oldNode = getNode(node.id) as IGatsbyNode
 
   // Ensure the plugin isn't creating a node type owned by another
   // plugin. Type "ownership" is first come first served.
@@ -825,7 +825,7 @@ export const createNode = (...args: Parameters<CreateNode>) => (
  * updated but still exist so Gatsby knows to keep them.
  */
 export const touchNode = (
-  options: string | { nodeId: IReduxNode["id"] },
+  options: string | { nodeId: IGatsbyNode["id"] },
   plugin?: IGatsbyPlugin
 ): ITouchNodeAction => {
   let nodeId = _.get(options, `nodeId`)
@@ -843,7 +843,7 @@ export const touchNode = (
     nodeId = options
   }
 
-  const node = getNode(nodeId) as IReduxNode
+  const node = getNode(nodeId) as IGatsbyNode
   if (node && !typeOwners[node.internal.type]) {
     typeOwners[node.internal.type] = node.internal.owner
   }
@@ -856,7 +856,7 @@ export const touchNode = (
 }
 
 interface ICreateNodeInput {
-  node: IReduxNode
+  node: IGatsbyNode
   fieldName?: string
   fieldValue?: string
   name?: string
@@ -944,7 +944,7 @@ export const createNodeField = (
  * instead.
  */
 export const createParentChildLink = (
-  { parent, child }: { parent: IReduxNode; child: IReduxNode },
+  { parent, child }: { parent: IGatsbyNode; child: IGatsbyNode },
   plugin?: IGatsbyPlugin
 ): ICreateParentChildLinkAction => {
   if (!parent.children.includes(child.id)) {
@@ -1233,7 +1233,30 @@ const maybeAddPathPrefix = (path: string, pathPrefix: string): string => {
  * of the box. You must have a plugin setup to integrate the redirect data with
  * your hosting technology e.g. the [Netlify
  * plugin](/packages/gatsby-plugin-netlify/), or the [Amazon S3
+<<<<<<< HEAD:packages/gatsby/src/redux/actions/public.ts
  * plugin](/packages/gatsby-plugin-s3/).
+=======
+ * plugin](/packages/gatsby-plugin-s3/). Alternatively, you can use
+ * [this plugin](/packages/gatsby-plugin-meta-redirect/) to generate meta redirect
+ * html files for redirecting on any static file host.
+ *
+ * @param {Object} redirect Redirect data
+ * @param {string} redirect.fromPath Any valid URL. Must start with a forward slash
+ * @param {boolean} redirect.isPermanent This is a permanent redirect; defaults to temporary
+ * @param {string} redirect.toPath URL of a created page (see `createPage`)
+ * @param {boolean} redirect.redirectInBrowser Redirects are generally for redirecting legacy URLs to their new configuration. If you can't update your UI for some reason, set `redirectInBrowser` to true and Gatsby will handle redirecting in the client as well.
+ * @param {boolean} redirect.force (Plugin-specific) Will trigger the redirect even if the `fromPath` matches a piece of content. This is not part of the Gatsby API, but implemented by (some) plugins that configure hosting provider redirects
+ * @param {number} redirect.statusCode (Plugin-specific) Manually set the HTTP status code. This allows you to create a rewrite (status code 200) or custom error page (status code 404). Note that this will override the `isPermanent` option which also sets the status code. This is not part of the Gatsby API, but implemented by (some) plugins that configure hosting provider redirects
+ * @example
+ * // Generally you create redirects while creating pages.
+ * exports.createPages = ({ graphql, actions }) => {
+ *   const { createRedirect } = actions
+ *   createRedirect({ fromPath: '/old-url', toPath: '/new-url', isPermanent: true })
+ *   createRedirect({ fromPath: '/url', toPath: '/zn-CH/url', Language: 'zn' })
+ *   createRedirect({ fromPath: '/not_so-pretty_url', toPath: '/pretty/url', statusCode: 200 })
+ *   // Create pages here
+ * }
+>>>>>>> upstream/master:packages/gatsby/src/redux/actions/public.js
  */
 export const createRedirect = ({
   fromPath,