+++ packages/gatsby/src/utils/babel-parse-to-ast.ts
@@ -1,0 +1,61 @@
+import { parse, ParserOptions } from "@babel/parser"
+import { File } from "@babel/types"
+
+const PARSER_OPTIONS: ParserOptions = {
+  allowImportExportEverywhere: true,
+  allowReturnOutsideFunction: true,
+  allowSuperOutsideMethod: true,
+  sourceType: `unambiguous`,
+  plugins: [
+    `jsx`,
+    `flow`,
+    `doExpressions`,
+    `objectRestSpread`,
+    [
+      `decorators`,
+      {
+        decoratorsBeforeExport: true,
+      },
+    ],
+    `classProperties`,
+    `classPrivateProperties`,
+    `classPrivateMethods`,
+    `exportDefaultFrom`,
+    `exportNamespaceFrom`,
+    `asyncGenerators`,
+    `functionBind`,
+    `functionSent`,
+    `dynamicImport`,
+    `numericSeparator`,
+    `optionalChaining`,
+    `importMeta`,
+    `bigInt`,
+    `optionalCatchBinding`,
+    `throwExpressions`,
+    [
+      `pipelineOperator`,
+      {
+        proposal: `minimal`,
+      },
+    ],
+    `nullishCoalescingOperator`,
+  ],
+}
+
+export function getBabelParserOptions(filePath: string): ParserOptions {
+  // Flow and TypeScript plugins can't be enabled simultaneously
+  if (/\.tsx?/.test(filePath)) {
+    const { plugins } = PARSER_OPTIONS
+    return {
+      ...PARSER_OPTIONS,
+      plugins: plugins!.map(plugin =>
+        plugin === `flow` ? `typescript` : plugin
+      ),
+    }
+  }
+  return PARSER_OPTIONS
+}
+
+export function babelParseToAst(contents: string, filePath: string): File {
+  return parse(contents, getBabelParserOptions(filePath))
+}